+++ Changelog.md
@@ -3,7 +3,7 @@
 ## 0.1.*
 
 ### 0.1.0
- - + Added `sendLocation`
+ - + Added support for Locations, Contacts, Files, Venues, Documents, Voice, Audio
  - + Added specific response classes
  - * Changed one simple format of Response to multiple classes `Text, Location, Keyboard, Photo ...`
  - * Changed format of **Response** descriptions